requirements.txt
html2text
pyyaml
requests
tqdm
threading
